refactor out schemaSet into a separate file to avoid confusing jsdoc-to-markdown
